feat: BookingForm with loading spinner
